quant-agent-trader/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quant-agent-trader"
version = "0.1.0"
description = "AI-powered quantitative trading system for Indian stock markets"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["trading", "quantitative", "stocks", "algorithmic-trading", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "yfinance>=0.2.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=24.2.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "bandit>=1.7.8",
    "safety>=3.0.0",
    "pip-audit>=3.0.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
quant-agent = "main:main_async"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "config*", "data*", "features*", "learning*", "models*", "portfolio*", "signals*", "strategies*", "utils*", "backtesting*", "execution*", "risk*", "ingestion*", "orchestration*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/fixtures
)/
'''

[tool.black.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["S101", "B011"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "skip"
ignore_missing_imports = true
pretty = true

[[tool.mypy.overrides]]
module = "yfinance.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "lightgbm.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "xgboost.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sklearn.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--disable-warnings",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "api: marks tests that require API calls",
]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "venv/*",
    ".venv/*",
    "*/tests/*",
    "*/migrations/*",
    "setup.py",
    "*/site-packages/*",
]

[tool.coverage.report]
precision = 2
skip_empty = true
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
